添加了保活

+++ app/src/main/java/com/guowei/draw/simpleweather/bean/CaiRealTimeBean.java
@@ -249,7 +249,7 @@ public class CaiRealTimeBean {
                  */
 
                 private String status;
-                private int intensity;
+                private double intensity;
                 private String datasource;
 
                 public String getStatus() {
@@ -260,11 +260,11 @@ public class CaiRealTimeBean {
                     this.status = status;
                 }
 
-                public int getIntensity() {
+                public double getIntensity() {
                     return intensity;
                 }
 
-                public void setIntensity(int intensity) {
+                public void setIntensity(double intensity) {
                     this.intensity = intensity;
                 }
